+++ core/internal/features_test.go
@@ -31,6 +31,7 @@ import (
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap/zapcore"
+	"go.uber.org/zap/zaptest/observer"
 	"gopkg.in/guregu/null.v4"
 
 	ocrcommontypes "github.com/smartcontractkit/libocr/commontypes"
@@ -451,14 +452,10 @@ func TestIntegration_DirectRequest(t *testing.T) {
 	}
 }
 
-func TestIntegration_AsyncEthTxRevertingTransaction(t *testing.T) {
-	cfg := cltest.NewTestGeneralConfig(t)
-	cfg.Overrides.SetTriggerFallbackDBPollInterval(100 * time.Millisecond)
-	operatorContracts := setupOperatorContracts(t)
+func setupAppForEthTx(t *testing.T, cfg *configtest.TestGeneralConfig, operatorContracts OperatorContracts) (app *cltest.TestApplication, sendingAddress common.Address, o *observer.ObservedLogs) {
 	b := operatorContracts.sim
-
 	lggr, o := logger.TestLoggerObserved(t, zapcore.DebugLevel)
-	app := cltest.NewApplicationWithConfigAndKeyOnSimulatedBlockchain(t, cfg, b, lggr)
+	app = cltest.NewApplicationWithConfigAndKeyOnSimulatedBlockchain(t, cfg, b, lggr)
 	b.Commit()
 
 	sendingKeys, err := app.KeyStore.Eth().SendingKeys(nil)
@@ -477,55 +474,144 @@ func TestIntegration_AsyncEthTxRevertingTransaction(t *testing.T) {
 	err = app.Start(testutils.Context(t))
 	require.NoError(t, err)
 
-	// b.Commit()
-	fmt.Println("BALLS 1")
-	time.Sleep(5 * time.Second)
-	for _, msg := range o.All() {
-		fmt.Println(msg)
-	}
 	testutils.WaitForLogMessage(t, o, "Subscribing to new heads on chain 1337")
+	testutils.WaitForLogMessage(t, o, "Subscribed to heads on chain 1337")
 
-	tomlSpec := `
+	return app, sendingKeys[0].Address.Address(), o
+}
+
+func TestIntegration_AsyncEthTx(t *testing.T) {
+	cfg := cltest.NewTestGeneralConfig(t)
+	cfg.Overrides.SetTriggerFallbackDBPollInterval(100 * time.Millisecond)
+	operatorContracts := setupOperatorContracts(t)
+	b := operatorContracts.sim
+
+	t.Run("with FailOnRevert enabled, run succeeds when transaction is successful", func(t *testing.T) {
+		app, sendingAddr, o := setupAppForEthTx(t, cfg, operatorContracts)
+		tomlSpec := `
 type            = "webhook"
 schemaVersion   = 1
 observationSource   = """
 	submit_tx  [type=ethtx to="%s"
             data="%s"
             minConfirmations="2"
+			failOnRevert=false
             from="[\\"%s\\"]"
 			]
 """
 `
-	// This data is a call to link token's `transfer` function and will revert due to insufficient LINK on the sender address
-	revertingData := "0xa9059cbb000000000000000000000000526485b5abdd8ae9c6a63548e0215a83e7135e6100000000000000000000000000000000000000000000000db069932ea4fe1400"
-	tomlSpec = fmt.Sprintf(tomlSpec, operatorContracts.linkTokenAddress.String(), revertingData, sendingKeys[0].Address.Address())
-	j := cltest.CreateJobViaWeb(t, app, []byte(cltest.MustJSONMarshal(t, web.CreateJobRequest{TOML: tomlSpec})))
-	cltest.AwaitJobActive(t, app.JobSpawner(), j.ID, testutils.WaitTimeout(t))
+		// This succeeds for whatever reason
+		revertingData := "0xdeadbeef"
+		tomlSpec = fmt.Sprintf(tomlSpec, operatorContracts.linkTokenAddress.String(), revertingData, sendingAddr)
+		j := cltest.CreateJobViaWeb(t, app, []byte(cltest.MustJSONMarshal(t, web.CreateJobRequest{TOML: tomlSpec})))
+		cltest.AwaitJobActive(t, app.JobSpawner(), j.ID, testutils.WaitTimeout(t))
 
-	fmt.Println("BALLS 3")
+		run := cltest.CreateJobRunViaUser(t, app, j.ExternalJobID, "")
+		assert.Equal(t, []*string([]*string(nil)), run.Outputs)
+		assert.Equal(t, []*string([]*string(nil)), run.Errors)
 
-	run := cltest.CreateJobRunViaUser(t, app, j.ExternalJobID, "")
-	assert.Equal(t, []*string([]*string(nil)), run.Outputs)
-	assert.Equal(t, []*string([]*string(nil)), run.Errors)
+		testutils.WaitForLogMessage(t, o, "Sent transaction")
+		b.Commit() // Needs at least two confirmations
+		b.Commit() // Needs at least two confirmations
+		b.Commit() // Needs at least two confirmations
+		testutils.WaitForLogMessage(t, o, "Resume run success")
 
-	fmt.Println("BALLS 4")
+		pipelineRuns := cltest.WaitForPipelineComplete(t, 0, j.ID, 1, 1, app.JobORM(), testutils.WaitTimeout(t), 100*time.Millisecond)
 
-	testutils.WaitForLogMessage(t, o, "Sent transaction")
-	b.Commit() // Needs at least two confirmations
-	b.Commit() // Needs at least two confirmations
-	b.Commit() // Needs at least two confirmations
-	// testutils.WaitForLogMessage(t, o, "Resume run success")
-	time.Sleep(5 * time.Second)
-	// fmt.Println("BALLS", o.All())
+		// The run should have succeeded but with the receipt detailing the reverted transaction
+		pipelineRun := pipelineRuns[0]
+		cltest.AssertPipelineTaskRunsSuccessful(t, pipelineRun.PipelineTaskRuns)
 
-	fmt.Println("BALLS 5")
+		outputs := pipelineRun.Outputs.Val.([]interface{})
+		require.Len(t, outputs, 1)
+		output := outputs[0]
+		receipt := output.(map[string]interface{})
+		assert.Equal(t, "0xc", receipt["blockNumber"])
+		assert.Equal(t, "0x7a120", receipt["gasUsed"])
+		assert.Equal(t, "0x1", receipt["status"])
+	})
 
-	// pipelineRuns := cltest.WaitForPipelineComplete(t, 0, j.ID, 1, 1, app.JobORM(), testutils.WaitTimeout(t), 100*time.Millisecond)
-	pipelineRuns := cltest.WaitForPipelineError(t, 0, j.ID, 1, 1, app.JobORM(), 5*time.Second, 100*time.Millisecond)
+	t.Run("with FailOnRevert enabled, run fails with transaction reverted error", func(t *testing.T) {
+		app, sendingAddr, o := setupAppForEthTx(t, cfg, operatorContracts)
+		tomlSpec := `
+type            = "webhook"
+schemaVersion   = 1
+observationSource   = """
+	submit_tx  [type=ethtx to="%s"
+            data="%s"
+            minConfirmations="2"
+			failOnRevert=true
+            from="[\\"%s\\"]"
+			]
+"""
+`
+		// This data is a call to link token's `transfer` function and will revert due to insufficient LINK on the sender address
+		revertingData := "0xa9059cbb000000000000000000000000526485b5abdd8ae9c6a63548e0215a83e7135e6100000000000000000000000000000000000000000000000db069932ea4fe1400"
+		tomlSpec = fmt.Sprintf(tomlSpec, operatorContracts.linkTokenAddress.String(), revertingData, sendingAddr)
+		j := cltest.CreateJobViaWeb(t, app, []byte(cltest.MustJSONMarshal(t, web.CreateJobRequest{TOML: tomlSpec})))
+		cltest.AwaitJobActive(t, app.JobSpawner(), j.ID, testutils.WaitTimeout(t))
 
-	// The run should have failed as a revert
-	pipelineRun := pipelineRuns[0]
-	cltest.AssertPipelineTaskRunsErrored(t, pipelineRun.PipelineTaskRuns)
+		run := cltest.CreateJobRunViaUser(t, app, j.ExternalJobID, "")
+		assert.Equal(t, []*string([]*string(nil)), run.Outputs)
+		assert.Equal(t, []*string([]*string(nil)), run.Errors)
+
+		testutils.WaitForLogMessage(t, o, "Sent transaction")
+		b.Commit() // Needs at least two confirmations
+		b.Commit() // Needs at least two confirmations
+		b.Commit() // Needs at least two confirmations
+		testutils.WaitForLogMessage(t, o, "Resume run success")
+
+		pipelineRuns := cltest.WaitForPipelineError(t, 0, j.ID, 1, 1, app.JobORM(), testutils.WaitTimeout(t), 100*time.Millisecond)
+
+		// The run should have failed as a revert
+		pipelineRun := pipelineRuns[0]
+		cltest.AssertPipelineTaskRunsErrored(t, pipelineRun.PipelineTaskRuns)
+	})
+
+	t.Run("with FailOnRevert disabled, run succeeds with output being reverted receipt", func(t *testing.T) {
+		app, sendingAddr, o := setupAppForEthTx(t, cfg, operatorContracts)
+		tomlSpec := `
+type            = "webhook"
+schemaVersion   = 1
+observationSource   = """
+	submit_tx  [type=ethtx to="%s"
+            data="%s"
+            minConfirmations="2"
+			failOnRevert=false
+            from="[\\"%s\\"]"
+			]
+"""
+`
+		// This data is a call to link token's `transfer` function and will revert due to insufficient LINK on the sender address
+		revertingData := "0xa9059cbb000000000000000000000000526485b5abdd8ae9c6a63548e0215a83e7135e6100000000000000000000000000000000000000000000000db069932ea4fe1400"
+		tomlSpec = fmt.Sprintf(tomlSpec, operatorContracts.linkTokenAddress.String(), revertingData, sendingAddr)
+		j := cltest.CreateJobViaWeb(t, app, []byte(cltest.MustJSONMarshal(t, web.CreateJobRequest{TOML: tomlSpec})))
+		cltest.AwaitJobActive(t, app.JobSpawner(), j.ID, testutils.WaitTimeout(t))
+
+		run := cltest.CreateJobRunViaUser(t, app, j.ExternalJobID, "")
+		assert.Equal(t, []*string([]*string(nil)), run.Outputs)
+		assert.Equal(t, []*string([]*string(nil)), run.Errors)
+
+		testutils.WaitForLogMessage(t, o, "Sent transaction")
+		b.Commit() // Needs at least two confirmations
+		b.Commit() // Needs at least two confirmations
+		b.Commit() // Needs at least two confirmations
+		testutils.WaitForLogMessage(t, o, "Resume run success")
+
+		pipelineRuns := cltest.WaitForPipelineComplete(t, 0, j.ID, 1, 1, app.JobORM(), testutils.WaitTimeout(t), 100*time.Millisecond)
+
+		// The run should have succeeded but with the receipt detailing the reverted transaction
+		pipelineRun := pipelineRuns[0]
+		cltest.AssertPipelineTaskRunsSuccessful(t, pipelineRun.PipelineTaskRuns)
+
+		outputs := pipelineRun.Outputs.Val.([]interface{})
+		require.Len(t, outputs, 1)
+		output := outputs[0]
+		receipt := output.(map[string]interface{})
+		assert.Equal(t, "0xc", receipt["blockNumber"])
+		assert.Equal(t, "0x7a120", receipt["gasUsed"])
+		assert.Equal(t, "0x0", receipt["status"])
+	})
 }
 
 func setupOCRContracts(t *testing.T) (*bind.TransactOpts, *backends.SimulatedBackend, common.Address, *offchainaggregator.OffchainAggregator, *flags_wrapper.Flags, common.Address) {